pyproject.toml
[project]
name = "github-license-analyzer"
version = "0.1.0"
description = "Analyzes license information for GitHub repositories"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "requests>=2.31.0",
    "tqdm>=4.66.0",
    "rapidfuzz>=3.0.0",
    "python-dotenv>=1.0.0",
    "google-generativeai>=0.8.5",
    "google-genai>=1.18.0",
    "pyyaml>=6.0.2",
    "packaging>=25.0",
    "beautifulsoup4>=4.13.4",
    "aiofiles>=24.1.0",
    "httpx>=0.28.1",
    "asyncio>=4.0.0",
    "tenacity>=9.1.2",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "openai>=1.106.1",
    "aiohttp>=3.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
